requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
greenlet
psycopg2-binary
pydantic
pydantic[email]
pydantic-settings
aiosqlite
python-jose
passlib[bcrypt]
httpx
pytest-asyncio
async-asgi-testclient
asyncpg
loguru
redis
